chore: core refactor and UI polish
