+++ src/font-size/editor.scss
@@ -5,6 +5,7 @@
 .block-editor-format-toolbar__blablablocks-font-size-popover {
 	.components-popover__content {
 		min-width: 20.625rem;
+		padding: 5px;
 	}
 
 	.reset-button {
